# Optimize MD5 performance